Harden folder preview batching and extract preview hook
Chunk oversize preview IPC, enforce an aggregate byte budget, clear stale visible paths on unmount, and move the preview workflow into useVisibleFilePreviews.
